Add Docker environment for iceoryx
Create a Docker image with a built iceoryx library,
RouDi daemon, and example applications.

Also provide a way to launch Docker containers in which
the RouDi daemon and example applications can be run.

Main work is done in the Dockerfile and the build_and_run.sh script.

The .screenrc file is just configuring screen to use bash as its shell.
This is convenient in case screen is used to run the daemon and examples.

The README.md documents the feature and suggested usage patterns.

Issues: iox-#15
Signed-off-by: Cornel Izbasa <badc0ded@gmail.com>