iox-#558 Using of makeSmartC, using custom deleter
Signed-off-by: Mueller Julian (CC-AD/EYF1) <julian.mueller@de.bosch.com>

iox-#558 fuzzingRouDiUDS now indicating if the connect was successful

Signed-off-by: Mueller Julian (CC-AD/EYF1) <julian.mueller@de.bosch.com>

iox-#558 Deleted comment

Signed-off-by: Mueller Julian (CC-AD/EYF1) <julian.mueller@de.bosch.com>

iox-#558 fuzzingRouDiUDS now closes socket directly after sending a message and returns a bool if the connection to the socket was successful. For com Fuzzing it is no longer checked if RouDi is running via UDS call.

Signed-off-by: Mueller Julian (CC-AD/EYF1) <julian.mueller@de.bosch.com>

iox-#558 Fixed a bug in fuzzingRouDiUDS where the sendto method was called even though the connection could not be created.

Signed-off-by: Mueller Julian (CC-AD/EYF1) <julian.mueller@de.bosch.com>
Signed-off-by: Hintz Martin (XC-AD/ESB5) <martin.hintz@de.bosch.com>