Allow cpptoml to be provided externally.
While adding IceOryx to Nixpkgs, I realized that the ExternalProject
mechanisms are somewhat problematic because they are not hermetic and
can fail easily depending on how strict the build sandbox is.

In this scenario, it is much easier to use `cpptoml` when provided
externally. The approach is inspired by `spdlog` and the `fmt` library.

There should be no change to the default build behavior.

In my use-case I'm adding IceOryx to the Nixpkgs eco-system, but this
would presumably also be useful for other packaging scenarios including
Bazel builds which use `rules_foreign_cc` to build IceOryx.