iox-#1194 Install hdrs to include/${PROJECT_NAME}
Adds a CMake option APPEND_PROJECT_NAME_TO_INCLUDEDIR defaulting to ON
that appends ${PROJECT_NAME} to the directory headers are installed.
This is to support ros2/ros2#1150.

See also: eclipse-cyclonedds/cyclonedds#1169

When using ROS 2 and Colcon it's possible to have two versions of a
package installed: typically one from the debian packages (like
ros-[rosdistro]-iceoryx-binding-c) and another in a from-source
workspace. In Colcon terms this is called "overriding a package", and
the /opt/ros/[rosdistro] folder would be called the underlay while the
from-source workspace would be called the overlay. A user would do this
when they want to try a newer version of a package without building
everything from source.

When two or more packages install their headers to the same directory
in an underlay, it becomes possible for packages in the overlay to
accidentally find the headers of an overridden package in the underlay.
This can cause build failures or undefined behavior at runtime
depending on the differences between the headers in the overridden and
overriding packages. The issue is avoided if a package installs its
headers to a unique folder using ${PROJECT_NAME}.

Signed-off-by: Shane Loretz <shane.loretz@gmail.com>