iox-#1311 Fix warning when building with CMake 3.23.
When building iceoryx_posh with CMake 3.23, it throws a warning:

Ignoring empty string ("") provided on the command line.

This happens because when CMAKE_TOOLCHAIN_FILE is not
defined, TOOLCHAIN_FILE is also not defined and thus is empty.
Workaround this new behavior by setting a variable that contains
a bunch of things, and may also *optionally* include the
define for the TOOLCHAIN file, which removes the warning.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>