iox-#2052 Refine test cases
Signed-off-by: Dennis Liu <dennis48161025@gmail.com>

The changes involved renaming the tests and ensuring the test content
better aligns with the ZOMBIE criteria. The primary object under test is
`SutComplex`. For each constructor or assignment, tests were conducted
on empty containers, single element containers, multiple elements
containers, and full capacity containers. In the case of moves,
additional checks were made to see if the `DTor` of `SutComplex` was
called (from `FixedPositionContainer::clear()`).