iox-#2052 Fix non-copyable class compile error
Signed-off-by: Dennis Liu <dennis48161025@gmail.com>

With the previous implementation of `copy_and_move_impl`, non-copyable
classes would encounter compilation failures due to non-compile-time
if-else branching. The compiler would check the deleted copy constructor
and assignment in the `else` branch, even though they would never be
called.

To address this, several helper structs have been implemented to change
runtime if-else branching to compile-time if-else branching for C++14.
If C++17 is used, we can simply employ `if constexpr` to make the
branching compile-time.

The test case `e1cc7c9f-c1b5-4047-811b-004302af5c00`
(UsingMoveCtorAtNonCopyableTypeShouldCompile) verifies that the
implementation works for both C++14 and C++17. The assignment portion
will be added in a subsequent commit.