iox-#2052 Add missing tests
Signed-off-by: Dennis Liu <dennis48161025@gmail.com>

Based on the discussion at
[https://github.com/eclipse-iceoryx/iceoryx/pull/2069#discussion_r1375494670],
additional tests have been incorporated.

Furthermore, some statistical results have been updated due to
modifications in the `copy_and_move_impl`. For instance, when a
non-empty container is moved to an empty container using the move
assignment, the `MoveCtor` is the function that is invoked, not the
`operator=`. This is because if the slot was previously free, we utilize
placement new to ensure proper functionality. Consequently, it is the
`MoveCtor` that gets called."