iox-#1129 Fix tests on Windows
This one was really funny. The variable where the runtime was stored was
created before RouDiEnv. At the end of the scope RouDiEnv got destructed
and also cleanup up the NamedPipe it created. The the runtime was destructed
and tried to use the previously destroyed NamedPipe to notify RouDiEnv
about its shutdown. This lead to a crash at the end of the scope.